% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyFMUtility.R
\name{apDeriv}
\alias{apDeriv}
\title{apDeriv}
\usage{
apDeriv(V, maxdeg, alpha)
}
\arguments{
\item{V}{An N x K matrix of the Slepian sequences (or another set of tapers)}

\item{maxdeg}{The highest degree polynomial at which to compute the associated polynomials}

\item{alpha}{A parameter of the Gegenbauer polynomials used as a starting point. The Gegenbauer polynomials are orthogonal on [-1,1] with the       respect to the weight function (1-x^2)^(alpha - 1/2)}
}
\value{
H The K x P matrix of eigencoefficients of the associated polynomials

G The N x P matrix of associated polynomials

D The N x P matrix of the derivatives of the associated polynomials

Hn The column sums of squares of R
}
\description{
Derives the associated polynomials and their derivatives.
}
\details{
The associated polynomials are created through a Gram-Schmidt type process involving the Slepian sequences and the Gegenbauer polynomials.
}
\examples{
 N <- 1024
 NW <- 4
 K <- 2*NW - 1
 DW <- dpss(N,K,NW,returnEigenvalues=FALSE)

 P <- 3

 apDeriv(DW$v,P,alpha=0.75)
}
\references{
D. J. Thomson, "Polynomial phase demodulation in multitaper analysis," 2009 IEEE/SP 15th Workshop on Statistical Signal Processing, 2009, pp. 401-404, doi: 10.1109/SSP.2009.5278553.

D. J. Thomson, "Inverse-constrained projection filters," Proc. SPIE 4478, Wavelets: Applications in Signal and Image Processing IX, (5 December 2001); https://doi.org/10.1117/12.449708
}
\author{
Created by Wesley Burr, edited slightly and expanded by Kian Blanchette
}
