% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyFMUtility.R
\name{computeEigencoefficients}
\alias{computeEigencoefficients}
\title{computeEigencoefficients}
\usage{
computeEigencoefficients(x, dw, nFFT)
}
\arguments{
\item{x}{A time series}

\item{dw}{An N x K matrix of orthogonal tapers}

\item{nFFT}{The length of the series after zero padding}
}
\value{
yk The K x nFFT matrix of eigencoefficients
}
\description{
Computes the eigencoefficients of a series given a set of tapers
}
\details{
The series is first zero padded and tapered. Then, the eigencoefficients are computed by Fast Fourier transform on the tapered, padded series.
}
\examples{
N <- 1024
NW <- 4
K <- 2*NW - 1
DW <- dpss(N,K,NW,returnEigenvalues=FALSE)

X <- rnorm(N)
nFFT <- 2^ceiling(log2(2*N))

yk <- computeEigencoefficients(X,DW$v, nFFT)
}
\references{
Thomson, David J. "Spectrum estimation and harmonic analysis." Proceedings of the IEEE 70.9 (1982): 1055-1096.
}
\author{
Kian Blanchette
}
