% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{dpssap}
\alias{dpssap}
\title{dpssap}
\usage{
dpssap(V,maxdeg,alpha=0.75)
}
\arguments{
\item{V}{An NxK matrix of Slepian sequences as created in the dpss function of the multitaper package.}

\item{maxdeg}{The highest degree of polynomial to be generated.}

\item{alpha}{A parameter of the Gegenbauer polynomials used as a starting point. The Gegenbauer polynomials are orthogonal on [-1,1] with the       respect to the weight function (1-x^2)^(alpha - 1/2)}
}
\value{
U The K x P matrix of eigencoefficients of the associated polynomials

R The N x P matrix of associated polynomials

Hn The column sums of squares of R
}
\description{
Computes a set of polynomials associated with the Slepian sequences. The polynomials themselves are not orthogonal, but when expanded in their eigencoeffiecent form U = V^T R, they are.
}
\details{
The associated polynomials are created through a Gram-Schmidt type process involving the Slepian sequences and the Gegenbauer polynomials.
}
\examples{
 N <- 1000
 NW <- 4
 K <- 2*NW - 1
 DW <- dpss(N,K,NW,returnEigenvalues=FALSE)

 P <- 3

 dpssap(DW$v,P,alpha=0.75)
}
\references{
D. J. Thomson, "Polynomial phase demodulation in multitaper analysis," 2009 IEEE/SP 15th Workshop on Statistical Signal Processing, 2009, pp. 401-404, doi: 10.1109/SSP.2009.5278553.

D. J. Thomson, "Inverse-constrained projection filters," Proc. SPIE 4478, Wavelets: Applications in Signal and Image Processing IX, (5 December 2001); https://doi.org/10.1117/12.449708
}
