% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{dpssDeriv}
\alias{dpssDeriv}
\title{dpssDeriv}
\usage{
dpssDeriv(DW,NW)
}
\arguments{
\item{DW}{The DPSS object as output by the dpss function of the multitaper package. This is essentially a list containing both the matrix of the first K Slepian sequences of length N and the associated eigenvalues. Both are needed here.}

\item{NW}{The multitaper time-bandwidth product (N = time = number of points * sampling period, W = bandwidth). This should have been selected already when computing the Slepian sequences. Typically, K = 2NW - 1, so a common choice of NW is (K+1)/2, where K is the number of tapers. The NW chosen here MUST match the NW used to generate the DW object.}
}
\value{
efnDeriv An N x K matrix whose columns contain the derivatives of the first K Slepian sequences.
}
\description{
Computes "derivatives" of the Slepian sequences at the same points the Slepians are defined at.
}
\details{
The derivatives of the Slepians are defined through a matrix equation involving a skew-symmetrix Toeplitz matrix and the Slepian sequences scaled by their eigenvalues. Thus, we only need one row of the Toeplitz matrix and we can use the Fast Fourier Transform and the Convolution Theorem to generate the derivatives as opposed to computing them directly through a lengthy matrix multiplication.
}
\examples{
N <- 1024
NW <- 4
K <- 2*NW - 1

DW <- dpss(N,K,NW,returnEigenvalues = TRUE)

Vdot <- dpssDeriv(DW,NW)
}
\references{
Blanchette, Kian. Multitaper statistical tests for the detection of frequency-modulated signals. MSc. Thesis. Queen's University (Canada), 2020.
}
