% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyFMUtility.R
\name{findLocalFMaxM}
\alias{findLocalFMaxM}
\alias{findlocalFMaxM}
\title{findLocalFMaxM}
\usage{
findLocalFMaxM(obj, k, cutoff)
}
\arguments{
\item{obj}{The F statistic object, a matrix of size (number of Fourier frequencies) x (number of polynomial degrees)}

\item{k}{The number of tapers used to compute the polynomial F statistics.}

\item{cutoff}{The threshold above which we want to find local maxima.}
}
\value{
MAXES A list of size P, where P is the number of polynomial degrees for which the F statistics were computed. Each element of the list contains the indices corresponding to the Fourier frequencies where local maxima can be found.
}
\description{
Finds local maxima that exceed a given threshold, typically a percentile of an F distribution.
}
\details{
The function finds all the indices that are above the provided cutoff. The local maxima are the turning points of this set of indices. That is, the local maxima are the points above the threshold that are greater than the points just before and just after them.
}
\examples{
N <- 1024
NW <- 4
K <- 2*NW - 1
stm11 <- 2.0/(N-1)
tt <- n * stm11 - 1.0

f <- 256 / N
n <- 0:(N-1)
poly <- 0.0025 * (1.0 - 2.0 * tt^2)
sig <- sin(2*pi*f*n + 2*pi*cumsum(poly))
noise <- rnorm(N)
X <- sig + noise

P <- 3
AP <- dpssap()
}
\references{
Blanchette, Kian. Multitaper statistical tests for the detection of frequency-modulated signals. MSc. Thesis. Queen's University (Canada), 2020.
}
\author{
Kian Blanchette, based on code written by Dave Riegert
}
