% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{IFcompute}
\alias{IFcompute}
\title{IFcompute}
\usage{
IFcompute(yk,V,Vdot)
}
\arguments{
\item{yk}{The matrix of eigencoefficients. M x K where M = number of Fourier frequencies, K = number of tapers.}

\item{V}{The matrix of Slepian sequences. N x K where N = length of the series in the time domains.}

\item{Vdot}{The matrix of derivatives of the Slepian sequences. N x K matrix.}
}
\value{
phi The N x M matrix of instantaneous frequency series, one column for each frequency at which the eigencoefficients are computed.
}
\description{
Computes the estimate of the instantaneous frequency of a time series based on the Slepian sequences.
}
\details{
For each frequency at which the eigencoefficients are computed, an inverse series and its derivative series are also computed. These are used together to produce a matrix of instantaneous frequency series in the time domain.
}
\examples{
N <- 1024
NW <- 4
K <- 2*NW - 1
DW <- dpss(N,K,NW,returnEigenvalues=FALSE)

Vdot <- dpssDeriv(DW,NW)

X <- rnorm(N)
sp <- spec.mtm(X, dpssIN = DW, returnInternals = TRUE)
yk <- sp$mtm$eigenCoefs

instFreq <- IFcompute(yk, DW$v, Vdot)
}
\references{
D. J. Thomson, "Inverse-constrained projection filters," Proc. SPIE 4478, Wavelets: Applications in Signal and Image Processing IX, (5 December 2001); https://doi.org/10.1117/12.449708

Blanchette, Kian. Multitaper statistical tests for the detection of frequency-modulated signals. MSc. Thesis. Queen's University (Canada), 2020.
}
