% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyFMUtility.R
\name{is.ap}
\alias{is.ap}
\title{is.ap}
\usage{
is.ap(obj)
}
\arguments{
\item{obj}{An R object}
}
\description{
A Boolean function that checks if an object is an associated polynomial of the Slepian sequences
}
\examples{
N <- 1000
NW <- 4
K <- 2*NW - 1
DW <- dpss(N,K,NW,returnEigenvalues=FALSE)

P <- 3

AP <- dpssap(DW$v,P,alpha=0.75)
is.ap(AP)
is.ap(DW)
}
\author{
Kian Blanchette
}
