% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyFMUtility.R
\name{is.deriv}
\alias{is.deriv}
\title{is.deriv}
\usage{
is.deriv(obj)
}
\arguments{
\item{obj}{An R object}
}
\description{
A Boolean function that checks if an object is a derivative of the Slepian sequences
}
\examples{
N <- 1024
NW <- 4
K <- 2*NW - 1

DW <- dpss(N,K,NW,returnEigenvalues = TRUE)

Vdot <- dpssDeriv(DW,NW)

is.deriv(Vdot)
is.deriv(DW)
}
\author{
Kian Blanchette
}
