% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyFMUtility.R
\name{is.dpss}
\alias{is.dpss}
\title{is.dpss}
\usage{
is.dpss(obj)
}
\arguments{
\item{obj}{An R object}
}
\description{
A Boolean function that checks if an object is a DPSS object
}
\examples{
N <- 1024
NW <- 4
K <- 2*NW - 1

DW <- dpss(N,K,NW,returnEigenvalues = TRUE)

Vdot <- dpssDeriv(DW,NW)

is.dpss(Vdot)
is.dpss(DW)
}
\author{
Kian Blanchette
}
