% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyFMUtility.R
\name{modF}
\alias{modF}
\title{modF}
\usage{
modF(mxdeg, nord, FPcoef, Fcoef, nfreqs)
}
\arguments{
\item{mxdeg}{Highest degree to compute the test statistics at}

\item{nord}{Number of tapers used}

\item{FPcoef}{Polynomial eigencoefficients of the instantaneous frequency series}

\item{Fcoef}{Zero-frequency eigencoefficients of the instantaneous frequency series}

\item{nfreqs}{Number of Fourier frequencies}
}
\value{
Fp Matrix of polynomial modulated test statistics
}
\description{
Fortan subroutine that computes the polynomial modulated test statistics
}
\details{
Calls a Fortran subroutine that loops over frequencies and polynomial degrees to compute the polynomial modulated test statistics
}
\examples{
N <- 1024
nFFT <- 2^ceiling(log2(2*N))
NW <- 4
K <- 2*NW - 1
DW <- dpss(N,K,NW,returnEigenvalues=FALSE)

X <- rnorm(N)
xzp <- c(X, rep(0,nFFT-N))
V <- rbind(DW$v, matrix(0, nrow = nFFT - N, ncol = K))
yk <- mvfft(V * X)

HF <- HarmonicF(yk, DW$v)
}
\references{
Blanchette, Kian. Multitaper statistical tests for the detection of frequency-modulated signals. MSc. Thesis. Queen's University (Canada), 2020.
}
\author{
Kian Blanchette
}
