% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyFMUtility.R
\name{ModulatedF}
\alias{ModulatedF}
\title{ModulatedF}
\usage{
ModulatedF(x, K, NW, P, alpha, nFFT, dpssIN, apIN, derivIN)
}
\arguments{
\item{x}{A time series}

\item{K}{The number of tapers}

\item{NW}{The multitaper time-bandwidth product}

\item{P}{The highest degree at which to compute the test statistics}

\item{alpha}{Parameter of the Gegenbauer polynomials used to generate the set of polynomials associated to the Slepian sequences. 0.75 by default}

\item{nFFT}{The length of the series after zero padding. 2^ceiling(log2(2*length(x))) by default}

\item{dpssIN}{A dpss (Slepian sequence) object that can be passed in instead of being generated within the function.}

\item{apIN}{An associated polynomial object that can be passed in instead of being generated within the function.}

\item{derivIN}{A dpss derivative object that can be passed in instead of being generated within the function.}

\item{returnInternals}{A Boolean object. If TRUE, return the intermediate objects created inside this function. If FALSE (the default), only return the test statistics.}
}
\value{
ModF The P x nFFT matrix containing the multitaper polynomial modulated test statistics.
}
\description{
Computes the multitaper polynomial modulated test statistics for a given time series
}
\details{
The estimates of the instantaneous frequency are computed for each Fourier frequency. Then the eigencoefficients and polynomial coefficients are derived for the instantaneous frequencies, which are then used to compute an F statistic.
}
\examples{
N <- 1024
NW <- 4
K <- 2*NW - 1
P <- 3
dpssIN <- dpss(N,K,NW,returnEigenvalues=FALSE)
apIN <- apDeriv(V = dpssIN$v, maxdeg = P, alpha = 0.75)
derivIN <- dpssDeriv(DW = dpssIN, NW = NW)

X <- rnorm(N)

mF <- ModulatedF(x, K, NW, P, alpha = 0.75, dpssIN = dpssIN, apIN = apIN, derivIN = derivIN)
}
\references{
Blanchette, Kian. Multitaper statistical tests for the detection of frequency-modulated signals. MSc. Thesis. Queen's University (Canada), 2020.
}
\author{
Kian Blanchette
}
